// backend/src/rag/dto/output/rag-steam-chunk-response.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { SimilarChunkResponseDto } from './similar-chunk-response.dto';

export class RagStreamChunkResponseDto {
  @ApiProperty({
    description: 'The text chunk generated by the LLM (can be empty)',
    example: ' NestJS',
  })
  content: string;

  @ApiProperty({
    description: 'indicates if the generation is done',
    example: false,
  })
  done: boolean;

  @ApiPropertyOptional({
    description: 'Sources cited by the model (only present in the last chunk)',
    type: [SimilarChunkResponseDto],
  })
  sources?: SimilarChunkResponseDto[];
}